[Backend] - PBI20

+++ service-task-SJ-3/src/main/java/sit/int221/servicetasksj3/controller/BoardController.java
@@ -5,25 +5,14 @@ import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
+import org.springframework.http.*;
 import org.springframework.web.bind.annotation.*;
-import sit.int221.servicetasksj3.dtos.boardsDTO.BoardRequestDTO;
-import sit.int221.servicetasksj3.dtos.boardsDTO.BoardResponseDTO;
+import sit.int221.servicetasksj3.dtos.boardsDTO.*;
 import sit.int221.servicetasksj3.dtos.limitsDTO.SimpleLimitDTO;
-import sit.int221.servicetasksj3.dtos.statusesDTO.StatusDTO;
-import sit.int221.servicetasksj3.dtos.statusesDTO.StatusDTOTwo;
-import sit.int221.servicetasksj3.dtos.tasksDTO.SimpleTaskDTO;
-import sit.int221.servicetasksj3.dtos.tasksDTO.TaskDTO;
-import sit.int221.servicetasksj3.dtos.tasksDTO.TaskDTOTwo;
-import sit.int221.servicetasksj3.dtos.tasksDTO.TaskNewDTO;
-import sit.int221.servicetasksj3.entities.Board;
-import sit.int221.servicetasksj3.entities.Task;
-import sit.int221.servicetasksj3.entities.TaskLimit;
-import sit.int221.servicetasksj3.entities.TaskStatus;
-import sit.int221.servicetasksj3.services.BoardService;
-import sit.int221.servicetasksj3.services.StatusService;
-import sit.int221.servicetasksj3.services.TaskService;
+import sit.int221.servicetasksj3.dtos.statusesDTO.*;
+import sit.int221.servicetasksj3.dtos.tasksDTO.*;
+import sit.int221.servicetasksj3.entities.*;
+import sit.int221.servicetasksj3.services.*;
 
 import java.util.List;
 
@@ -41,15 +30,19 @@ public class BoardController {
     @Autowired
     private ModelMapper modelMapper;
 
+    // false - read operation (get)
+    // true - write operation (create, update, delete)
 
+    // Board
     @GetMapping("")
     public ResponseEntity<List<BoardResponseDTO>> getBoardIdByOwner() {
         List<BoardResponseDTO> boardIds = boardService.getBoardIdByOwner();
         return ResponseEntity.ok(boardIds);
     }
-    @GetMapping("/{id}")
-    public ResponseEntity<BoardResponseDTO> getBoardById(@PathVariable String id) {
-        BoardResponseDTO boardResponse = boardService.getBoardById(id);
+    @GetMapping("/{boardId}")
+    public ResponseEntity<BoardResponseDTO> getBoardById(@PathVariable String boardId) {
+        boardService.checkAuthorization(boardId, false);
+        BoardResponseDTO boardResponse = boardService.getBoardById(boardId);
         return ResponseEntity.ok(boardResponse);
     }
     @PostMapping("")
@@ -59,8 +52,21 @@ public class BoardController {
     }
     @DeleteMapping("/{boardId}")
     public ResponseEntity<Board> removeBoard(@PathVariable String boardId) {
+        boardService.checkAuthorization(boardId, true);
         Board deleted = boardService.removeBoard(boardId);
         return ResponseEntity.ok().body(deleted);
+    }
+    @PutMapping("/{boardId}")
+    public ResponseEntity<BoardResponseDTO> editBoard(@PathVariable String boardId, @Valid @RequestBody BoardRequestDTO boardRequest) {
+        boardService.checkAuthorization(boardId, true);
+        BoardResponseDTO updatedBoard = boardService.editBoard(boardId, boardRequest);
+        return ResponseEntity.ok(updatedBoard);
+    }
+    @PatchMapping("/{boardId}")
+    public ResponseEntity<VisibilityDTO> editBoardVisibility(@PathVariable String boardId, @Valid @RequestBody VisibilityDTO visibility) {
+        boardService.checkAuthorization(boardId, true);
+        VisibilityDTO updatedVisibility = boardService.editBoardVisibility(boardId, visibility);
+        return ResponseEntity.ok(updatedVisibility);
     }
 
     // Task of board
@@ -70,28 +76,33 @@ public class BoardController {
             @RequestParam(required = false, defaultValue = "id") String sortBy,
             @RequestParam(required = false) String[] filterStatuses
     ) {
-        boardService.validateBoardExists(boardId);
+//        boardService.validateBoardExists(boardId);
+        boardService.checkAuthorization(boardId, false);
         return taskService.getAllTasksFiltered(boardId, sortBy, filterStatuses);
     }
     @GetMapping("/{boardId}/tasks/{taskId}")
     public SimpleTaskDTO getTaskById(@PathVariable String boardId,@PathVariable Integer taskId) {
+        boardService.checkAuthorization(boardId, false);
         Task task = taskService.findByID(boardId, taskId);
         SimpleTaskDTO simpleTaskDTO = modelMapper.map(task, SimpleTaskDTO.class);
         return simpleTaskDTO;
     }
     @PostMapping("/{boardId}/tasks")
     public ResponseEntity<TaskDTOTwo> createNewTasks(@PathVariable String boardId, @Valid @RequestBody TaskNewDTO task) {
+        boardService.checkAuthorization(boardId, true);
         Task createTask = taskService.createNewTasks(boardId, task);
         TaskDTOTwo createdTaskDTO = modelMapper.map(createTask, TaskDTOTwo.class);
         return ResponseEntity.status(HttpStatus.CREATED).body(createdTaskDTO);
     }
     @DeleteMapping("/{boardId}/tasks/{taskId}")
     public ResponseEntity<TaskDTO> removeTasks(@PathVariable String boardId,@PathVariable Integer taskId) {
+        boardService.checkAuthorization(boardId, true);
         TaskDTO deletedTaskDTO = taskService.removeTasks(boardId, taskId);
         return ResponseEntity.ok().body(deletedTaskDTO);
     }
     @PutMapping("/{boardId}/tasks/{taskId}")
     public ResponseEntity<TaskDTOTwo> updateTasks(@PathVariable String boardId, @PathVariable Integer taskId, @Valid @RequestBody TaskNewDTO task) {
+        boardService.checkAuthorization(boardId, true);
         Task updatedTask = taskService.updateTask(boardId, taskId, task);
         TaskDTOTwo updatedTaskDTO = modelMapper.map(updatedTask, TaskDTOTwo.class);
         return ResponseEntity.ok(updatedTaskDTO);
@@ -99,33 +110,40 @@ public class BoardController {
 
     // Statuses of board
     @GetMapping("/{boardId}/statuses")
-    public List<StatusDTOTwo> getAllStatuses(@PathVariable String boardId) {
-        boardService.validateBoardExists(boardId);
-        return statusService.getAllStatuses(boardId);
+    public ResponseEntity<List<StatusDTOTwo>> getAllStatuses(@PathVariable String boardId) {
+//        boardService.validateBoardExists(boardId);
+        boardService.checkAuthorization(boardId, false);
+        return ResponseEntity.ok(statusService.getAllStatuses(boardId));
     }
     @GetMapping("/{boardId}/statuses/{statusId}")
     public TaskStatus getStatusesById(@PathVariable String boardId, @PathVariable Integer statusId){
+        boardService.checkAuthorization(boardId, false);
         return statusService.getStatusesById(boardId, statusId);
     }
     @GetMapping("/{boardId}/statuses/limit")
     public ResponseEntity<TaskLimit> getStatusesLimit(@PathVariable String boardId){
+        boardService.checkAuthorization(boardId, false);
         return ResponseEntity.ok(statusService.getStatusesLimit(boardId));
     }
     @PostMapping("/{boardId}/statuses")
     public ResponseEntity<StatusDTO> createNewStatuses(@PathVariable String boardId, @Valid @RequestBody StatusDTO status) {
+        boardService.checkAuthorization(boardId, true);
         StatusDTO createdStatus = statusService.createNewStatuses(boardId, status);
         return ResponseEntity.status(HttpStatus.CREATED).body(createdStatus);
     }
     @PutMapping("/{boardId}/statuses/{statusId}")
     public ResponseEntity<Object> updateStatuses(@PathVariable String boardId, @PathVariable Integer statusId, @RequestBody TaskStatus task) {
+        boardService.checkAuthorization(boardId, true);
         return ResponseEntity.ok(statusService.updateStatuses(boardId, statusId, task));
     }
     @DeleteMapping("/{boardId}/statuses/{statusId}")
     public ResponseEntity<Object> removeStatuses(@PathVariable String boardId,@Valid @PathVariable Integer statusId){
+        boardService.checkAuthorization(boardId, true);
         return ResponseEntity.ok(statusService.removeStatuses(boardId, statusId));
     }
     @DeleteMapping("/{boardId}/statuses/{statusId}/{newId}")
     public ResponseEntity<Object> removeStatusAndReplace(@PathVariable String boardId, @Valid @PathVariable Integer statusId , @PathVariable Integer newId){
+        boardService.checkAuthorization(boardId, true);
         return ResponseEntity.ok(statusService.transferStatuses(boardId,statusId,newId));
     }
     @PatchMapping("/{boardId}/statuses/maximumtask")
@@ -133,6 +151,7 @@ public class BoardController {
             @PathVariable String boardId,
             @RequestParam @Min(0) @Max(30) Integer maximumTask,
             @RequestParam Boolean isLimit) {
+        boardService.checkAuthorization(boardId, true);
         return ResponseEntity.ok(statusService.updateLimitTask(boardId, maximumTask, isLimit));
     }
 }
